+++ permission_handler_android/android/src/main/java/com/baseflow/permissionhandler/PermissionManager.java
@@ -24,6 +24,7 @@ import androidx.core.app.NotificationManagerCompat;
 import androidx.core.content.ContextCompat;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
@@ -72,12 +73,7 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
 
     @Override
     public boolean onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode != PermissionConstants.PERMISSION_CODE_IGNORE_BATTERY_OPTIMIZATIONS &&
-            requestCode != PermissionConstants.PERMISSION_CODE_MANAGE_EXTERNAL_STORAGE &&
-            requestCode != PermissionConstants.PERMISSION_CODE_SYSTEM_ALERT_WINDOW &&
-            requestCode != PermissionConstants.PERMISSION_CODE_REQUEST_INSTALL_PACKAGES &&
-            requestCode != PermissionConstants.PERMISSION_CODE_ACCESS_NOTIFICATION_POLICY &&
-            requestCode != PermissionConstants.PERMISSION_CODE_SCHEDULE_EXACT_ALARM) {
+        if (activity == null) {
             return false;
         }
 
@@ -144,7 +140,7 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
         pendingRequestCount--;
 
         // Post result if all requests have been handled.
-        if (pendingRequestCount == 0) {
+        if (successCallback != null && pendingRequestCount == 0) {
             this.successCallback.onSuccess(requestResults);
         }
         return true;
@@ -170,8 +166,37 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
             return false;
         }
 
+        // Calendar permissions are split between WRITE and READ in Android, and split between WRITE
+        // and FULL ACCESS in the plugin. We need special logic for this translation.
+        final List<String> permissionList = Arrays.asList(permissions);
+        final int calendarWriteIndex = permissionList.indexOf(Manifest.permission.WRITE_CALENDAR);
+        // WRITE -> WRITE.
+        if (calendarWriteIndex >= 0) {
+            final int writeGrantResult = grantResults[calendarWriteIndex];
+            final @PermissionConstants.PermissionStatus int writeStatus =
+                PermissionUtils.toPermissionStatus(this.activity, Manifest.permission.WRITE_CALENDAR, writeGrantResult);
+            requestResults.put(PermissionConstants.PERMISSION_GROUP_CALENDAR_WRITE_ONLY, writeStatus);
+
+            // WRITE + READ -> FULL ACCESS.
+            final int calendarReadIndex = permissionList.indexOf(Manifest.permission.READ_CALENDAR);
+            if (calendarReadIndex >= 0) {
+                final int readGrantResult = grantResults[calendarReadIndex];
+                final @PermissionConstants.PermissionStatus int readStatus =
+                    PermissionUtils.toPermissionStatus(this.activity, Manifest.permission.READ_CALENDAR, readGrantResult);
+                final @PermissionConstants.PermissionStatus int fullAccessStatus = strictestStatus(writeStatus, readStatus);
+                requestResults.put(PermissionConstants.PERMISSION_GROUP_CALENDAR_FULL_ACCESS, fullAccessStatus);
+                // Support deprecated CALENDAR permission.
+                requestResults.put(PermissionConstants.PERMISSION_GROUP_CALENDAR, fullAccessStatus);
+            }
+        }
+
         for (int i = 0; i < permissions.length; i++) {
             final String permissionName = permissions[i];
+
+            // WRITE_CALENDAR and READ_CALENDAR permission results have already been handled.
+            if (permissionName.equals(Manifest.permission.WRITE_CALENDAR) || permissionName.equals(Manifest.permission.READ_CALENDAR)) {
+                continue;
+            }
 
             @PermissionConstants.PermissionGroup final int permission =
                 PermissionUtils.parseManifestName(permissionName);
@@ -228,14 +253,12 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
                     permission,
                     PermissionUtils.toPermissionStatus(this.activity, permissionName, result));
             }
-
-            PermissionUtils.updatePermissionShouldShowStatus(this.activity, permission);
         }
 
         pendingRequestCount -= grantResults.length;
 
         // Post result if all requests have been handled.
-        if (pendingRequestCount == 0) {
+        if (successCallback != null && pendingRequestCount == 0) {
             this.successCallback.onSuccess(requestResults);
         }
         return true;
@@ -385,6 +408,16 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
                 launchSpecialPermission(
                     Settings.ACTION_REQUEST_SCHEDULE_EXACT_ALARM,
                     PermissionConstants.PERMISSION_CODE_SCHEDULE_EXACT_ALARM);
+            } else if (permission == PermissionConstants.PERMISSION_GROUP_CALENDAR_FULL_ACCESS || permission == PermissionConstants.PERMISSION_GROUP_CALENDAR) {
+                // Deny CALENDAR_FULL_ACCESS permission if manifest is not listing both write- and read permissions.
+                final boolean isValidManifest = isValidManifestForCalendarFullAccess();
+                if (isValidManifest) {
+                    permissionsToRequest.add(Manifest.permission.WRITE_CALENDAR);
+                    permissionsToRequest.add(Manifest.permission.READ_CALENDAR);
+                    pendingRequestCount += 2;
+                } else {
+                    requestResults.put(permission, PermissionConstants.PERMISSION_STATUS_DENIED);
+                }
             } else {
                 permissionsToRequest.addAll(names);
                 pendingRequestCount += names.size();
@@ -401,7 +434,7 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
         }
 
         // Post results immediately if no requests are pending.
-        if (pendingRequestCount == 0) {
+        if (this.successCallback != null && pendingRequestCount == 0) {
             this.successCallback.onSuccess(requestResults);
         }
     }
@@ -423,6 +456,13 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
             if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) {
                 return checkBluetoothPermissionStatus();
             }
+        }
+
+        // Inspect the manifest for CALENDAR_FULL_ACCESS, as a misconfigured manifest will give a false positive if READ access has been provided.
+        if (permission == PermissionConstants.PERMISSION_GROUP_CALENDAR_FULL_ACCESS || permission == PermissionConstants.PERMISSION_GROUP_CALENDAR) {
+            boolean isValidManifest = isValidManifestForCalendarFullAccess();
+            if (!isValidManifest)
+                return PermissionConstants.PERMISSION_STATUS_DENIED;
         }
 
         final List<String> names = PermissionUtils.getManifestNames(context, permission);
@@ -458,24 +498,19 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
                 : PermissionConstants.PERMISSION_STATUS_DENIED;
         }
 
-        final boolean targetsMOrHigher = context.getApplicationInfo().targetSdkVersion >= Build.VERSION_CODES.M;
+        final boolean requiresExplicitPermission = context.getApplicationInfo().targetSdkVersion >= Build.VERSION_CODES.M;
 
-        Set<@PermissionConstants.PermissionStatus Integer> permissionStatuses = new HashSet<>();
-        for (String name : names) {
-            // Only handle them if the client app actually targets a API level greater than M.
-            if (targetsMOrHigher) {
+        if (requiresExplicitPermission) {
+            final Set<@PermissionConstants.PermissionStatus Integer> permissionStatuses = new HashSet<>();
+
+            for (String name : names) {
                 if (permission == PermissionConstants.PERMISSION_GROUP_IGNORE_BATTERY_OPTIMIZATIONS) {
                     String packageName = context.getPackageName();
                     PowerManager pm = (PowerManager) context.getSystemService(Context.POWER_SERVICE);
-                    // PowerManager.isIgnoringBatteryOptimizations has been included in Android M first.
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                        if (pm != null && pm.isIgnoringBatteryOptimizations(packageName)) {
-                            permissionStatuses.add(PermissionConstants.PERMISSION_STATUS_GRANTED);
-                        } else {
-                            permissionStatuses.add(PermissionConstants.PERMISSION_STATUS_DENIED);
-                        }
+                    if (pm != null && pm.isIgnoringBatteryOptimizations(packageName)) {
+                        permissionStatuses.add(PermissionConstants.PERMISSION_STATUS_GRANTED);
                     } else {
-                        permissionStatuses.add(PermissionConstants.PERMISSION_STATUS_RESTRICTED);
+                        permissionStatuses.add(PermissionConstants.PERMISSION_STATUS_DENIED);
                     }
                 } else if (permission == PermissionConstants.PERMISSION_GROUP_MANAGE_EXTERNAL_STORAGE) {
                     if (Build.VERSION.SDK_INT < Build.VERSION_CODES.R) {
@@ -487,12 +522,10 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
                         : PermissionConstants.PERMISSION_STATUS_DENIED;
                     permissionStatuses.add(status);
                 } else if (permission == PermissionConstants.PERMISSION_GROUP_SYSTEM_ALERT_WINDOW) {
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                        int status = Settings.canDrawOverlays(context)
-                            ? PermissionConstants.PERMISSION_STATUS_GRANTED
-                            : PermissionConstants.PERMISSION_STATUS_DENIED;
-                        permissionStatuses.add(status);
-                    }
+                    int status = Settings.canDrawOverlays(context)
+                        ? PermissionConstants.PERMISSION_STATUS_GRANTED
+                        : PermissionConstants.PERMISSION_STATUS_DENIED;
+                    permissionStatuses.add(status);
                 } else if (permission == PermissionConstants.PERMISSION_GROUP_REQUEST_INSTALL_PACKAGES) {
                     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
                         int status = context.getPackageManager().canRequestPackageInstalls()
@@ -501,13 +534,11 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
                         permissionStatuses.add(status);
                     }
                 } else if (permission == PermissionConstants.PERMISSION_GROUP_ACCESS_NOTIFICATION_POLICY) {
-                    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                        NotificationManager notificationManager = (NotificationManager) context.getSystemService(Application.NOTIFICATION_SERVICE);
-                        int status = notificationManager.isNotificationPolicyAccessGranted()
-                            ? PermissionConstants.PERMISSION_STATUS_GRANTED
-                            : PermissionConstants.PERMISSION_STATUS_DENIED;
-                        permissionStatuses.add(status);
-                    }
+                    NotificationManager notificationManager = (NotificationManager) context.getSystemService(Application.NOTIFICATION_SERVICE);
+                    int status = notificationManager.isNotificationPolicyAccessGranted()
+                        ? PermissionConstants.PERMISSION_STATUS_GRANTED
+                        : PermissionConstants.PERMISSION_STATUS_DENIED;
+                    permissionStatuses.add(status);
                 } else if (permission == PermissionConstants.PERMISSION_GROUP_SCHEDULE_EXACT_ALARM) {
                     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
                         AlarmManager alarmManager = (AlarmManager) context.getSystemService(Context.ALARM_SERVICE);
@@ -525,11 +556,11 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
                     }
                 }
             }
+            if (!permissionStatuses.isEmpty()) {
+                return strictestStatus(permissionStatuses);
+            }
         }
 
-        if (!permissionStatuses.isEmpty()) {
-            return strictestStatus(permissionStatuses);
-        }
         return PermissionConstants.PERMISSION_STATUS_GRANTED;
     }
 
@@ -544,6 +575,10 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
      * @param requestCode      a request code to verify incoming results.
      */
     private void launchSpecialPermission(String permissionAction, int requestCode) {
+        if (activity == null) {
+            return;
+        }
+
         Intent intent = new Intent(permissionAction);
         if (!permissionAction.equals(Settings.ACTION_NOTIFICATION_POLICY_ACCESS_SETTINGS)) {
             String packageName = activity.getPackageName();
@@ -611,5 +646,24 @@ final class PermissionManager implements PluginRegistry.ActivityResultListener, 
             return PermissionConstants.PERMISSION_STATUS_DENIED;
         }
         return PermissionConstants.PERMISSION_STATUS_GRANTED;
+    }
+
+    /**
+     * Checks if the manifest contains both {@link Manifest.permission#WRITE_CALENDAR} and
+     * {@link Manifest.permission#READ_CALENDAR} permission declarations.
+     */
+    private boolean isValidManifestForCalendarFullAccess() {
+        List<String> names = PermissionUtils.getManifestNames(context, PermissionConstants.PERMISSION_GROUP_CALENDAR_FULL_ACCESS);
+        final boolean writeInManifest = names != null && names.contains(Manifest.permission.WRITE_CALENDAR);
+        final boolean readInManifest = names != null && names.contains(Manifest.permission.READ_CALENDAR);
+        final boolean validManifest = writeInManifest && readInManifest;
+        if (!validManifest) {
+            if (!writeInManifest)
+                Log.d(PermissionConstants.LOG_TAG, Manifest.permission.WRITE_CALENDAR + " missing in manifest");
+            if (!readInManifest)
+                Log.d(PermissionConstants.LOG_TAG, Manifest.permission.READ_CALENDAR + " missing in manifest");
+            return false;
+        }
+        return true;
     }
 }