+++ permission_handler_platform_interface/CHANGELOG.md
@@ -1,3 +1,16 @@
+## 4.0.1
+
+* Updates Android documentation on how to use `permission.photo` on Android 12 (API 32) and below and Android 13 (API 33) and above.
+
+## 4.0.0
+
+* **BREAKING CHANGE**: Replaces `Permission.calendarReadOnly` with `Permission.calendarWriteOnly`.
+
+## 3.12.0
+
+* Adds `Permission.calendarReadOnly` and `Permission.calendarFullAccess`.
+* Deprecates `Permission.calendar`. Developers should use `Permission.calendarReadOnly` and `Permission.calendarFullAccess` instead.
+
 ## 3.11.5
 
 * Updates the mentions of Android versions throughout the plugin, now following a format of 'Android {name} (API {number})'. For example: 'Android 13 (API 33)'.